Made average and std calculations more accurate

+++ benchmarks/src/main/scala/new_benchmarks/Processing.scala
@@ -2,14 +2,14 @@ package new_benchmarks
 
 import join_patterns.matching.MatchingAlgorithm
 
-private def standardDeviation(values: Seq[Long], average: Long): Double =
+private def standardDeviation(values: Seq[Long], average: Double): Double =
   Math.sqrt(values.map(v => Math.pow(v - average, 2)).sum / (values.size - 1))
 
-final case class ProcessedRepetitions(data: Seq[Long], average: Long, std: Option[Double])
+final case class ProcessedRepetitions(data: Seq[Long], average: Double, std: Option[Double])
 def processRepetitions(repetitions: Repetitions): ProcessedRepetitions =
   val millis = repetitions.map(_.toMillis)
 
-  val averageMillis = millis.sum / millis.size
+  val averageMillis = millis.sum.toDouble / millis.size
 
   if millis.size > 1 then ProcessedRepetitions(millis, averageMillis, Some(standardDeviation(millis, averageMillis)))
   else ProcessedRepetitions(millis, averageMillis, None)
